requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
pillow>=10.0.0
ffmpeg-python>=0.2.0
aiofiles>=23.0.0
jinja2>=3.1.0
